pyproject.toml
[project]
name = "matrix-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "matrix-nio>=0.25.2",
    "sqlalchemy>=2.0.39",
]
